refactor(Elevator): move states to package level, change various methods to package private
Temporarily changed attribute in context to public to make code run.
Plan to write a getter method to change attribute event back to protected access modifier.

Main() runs. No side effects.
